backend/requirements.txt
Django==2.2.8
django-filter==2.4.0
gunicorn==20.0.4